done patch of the survey form

+++ apps/candyland/src/app/engagements/new-survey/containers/new-survey/new-survey.component.ts
@@ -1,13 +1,13 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
-import { AbstractControl, FormArray, FormBuilder, FormControl, FormGroup, Validators } from '@angular/forms';
+import { AbstractControl, FormArray, FormControl, FormGroup } from '@angular/forms';
 import { moveItemInArray, transferArrayItem } from '@angular/cdk/drag-drop';
-import { Observable, Subject } from 'rxjs';
-import { debounceTime, tap, map } from 'rxjs/operators';
+import { combineLatest, Observable, of, Subject } from 'rxjs';
+import { debounceTime, tap, map, switchMap } from 'rxjs/operators';
 import { NewSurveyForm } from 'src/app/engagements/new-survey/new-survey-form';
 import { ControlsName } from '../../../../models/controls-name';
 import { AvailableNewEngagementService, RoutingStateService, SettingsService, SurveyService } from '@cl-core/services';
 import { QuestionFormFieldService } from '@cl-shared';
-import { Router } from '@angular/router';
+import { ActivatedRoute, ParamMap, Router } from '@angular/router';
 import { untilDestroyed } from 'ngx-take-until-destroy';
 import { ImageControlValue } from '@cl-helpers/image-control-value';
 import { Tenants } from '@cl-core/http-adapters/setting-json-adapter';
@@ -30,6 +30,14 @@ export class NewSurveyComponent implements OnInit, OnDestroy {
 
   public questionData$ = new Subject();
   public tenantSettings: ITenantsProperties;
+
+  constructor(private questionFormFieldService: QuestionFormFieldService,
+              private availableNewEngagementService: AvailableNewEngagementService,
+              private surveyService: SurveyService, private route: ActivatedRoute,
+              private router: Router, private routingState: RoutingStateService,
+              private cdr: ChangeDetectorRef, private settingsService: SettingsService) {
+  }
+
   // tslint:disable
   public get listId(): string {
     const id = this.questionFormFieldService.listId;
@@ -81,37 +89,66 @@ export class NewSurveyComponent implements OnInit, OnDestroy {
     return this.questionData$.asObservable();
   }
 
-  constructor(private questionFormFieldService: QuestionFormFieldService,
-    private availableNewEngagementService: AvailableNewEngagementService,
-    private surveyService: SurveyService,
-    private router: Router,
-    private routingState: RoutingStateService,
-    private cdr: ChangeDetectorRef,
-    private settingsService: SettingsService,
-              private fb: FormBuilder) {
-  }
-
   public ngOnInit(): void {
     this.initForm();
     this.getSurveyData();
     this.subscribeFormValueChanges();
     this.getTenants();
+
+    combineLatest([this.getSurveyData(), this.handleRouteParams()])
+      .pipe(untilDestroyed(this))
+      .subscribe(([questionData, question]) => {
+
+        // remove default value if we edit the existing question
+        if (question) {
+          this.deleteQuestion(0);
+        }
+
+        const patchData = question || this.getDefaultValue(questionData);
+
+        this.patchForm(patchData);
+        this.cdr.detectChanges();
+      });
   }
 
   public ngOnDestroy(): void {
   }
-  // TODO: need for the future patch form
-  public patchForm(): void {
-    this.deleteQuestion(0);
-    const data = NewSurveyForm.getDefaultValue();
+
+  public patchForm(data?): void {
+
+    // patch first simple fields fo the form
     this.form.patchValue(data, {emitEvent: false});
-    data.questions.forEach((item) => {
-      const group = this.createControlQuestion(item.payload.type);
-      console.log('question need to changes', item);
-      this.pathChoicePicture(item, group);
-      group.patchValue(item);
+
+    // patch other form fields
+    if (data.questions) {
+      this.patchQuestionGroups(data.questions);
+    }
+
+  }
+
+  public patchQuestionGroups(array: any[]): void {
+    array.forEach((item) => {
+      const group = this.groupPatchHandler(item);
       this.surveyQuestion.push(group);
     });
+  }
+
+  public patchGroup(item: IQuestion, mainGroup: FormGroup): void {
+    if (item.payload.type === SurveyQuestionType.questionGroup) {
+      item.payload.questions.forEach((item) => {
+        const group = this.groupPatchHandler(item);
+        (mainGroup.get('payload.questions') as FormArray).push(group);
+      });
+    }
+  }
+
+  public groupPatchHandler(item: IQuestion): FormGroup {
+    const group = this.createControlQuestion(item.payload.type);
+    this.pathChoicePicture(item, group);
+    this.patchMultipleChoice(item, group);
+    this.patchGroup(item, group);
+    group.patchValue(item);
+    return group;
   }
 
   public comeBack(): void {
@@ -127,16 +164,12 @@ export class NewSurveyComponent implements OnInit, OnDestroy {
   }
 
   public save(): void {
-    console.log(this.form.value);
-    // this.surveyService.createSurvey(this.form.value)
-    //   .pipe(
-    //     untilDestroyed(this),
-    //     map((engagement: IResponseApi<IEngagementApi>) => EngagementHttpAdapter.transformEngagement(engagement.data))
-    //   )
-    //   .subscribe((data: IEngagement) => {
-    //     this.availableNewEngagementService.setNewEngagement(data);
-    //     this.router.navigateByUrl('/engagements');
-    //   });
+    this.surveyService.createSurvey(this.form.value)
+      .pipe(untilDestroyed(this), map((engagement: IResponseApi<IEngagementApi>) => EngagementHttpAdapter.transformEngagement(engagement.data)))
+      .subscribe((data: IEngagement) => {
+        this.availableNewEngagementService.setNewEngagement(data);
+        this.router.navigateByUrl('/engagements');
+      });
   }
 
   public deleteQuestion(index: number) {
@@ -158,8 +191,7 @@ export class NewSurveyComponent implements OnInit, OnDestroy {
 
   private patchMultipleChoice(item: any, group: FormGroup): void {
     if (item.payload.type === SurveyQuestionType.multipleChoice) {
-      // TODO: need implement it
-      // this.questionFormFieldService.pathChoicePicture(item, group);
+      this.questionFormFieldService.patchMultipleChoice(item, group);
     }
   }
 
@@ -178,24 +210,15 @@ export class NewSurveyComponent implements OnInit, OnDestroy {
     this.addQuestion(SurveyQuestionType.rating);
   }
 
-  private getSurveyData(): void {
-    this.surveyData$ = this.surveyService.getSurveyData()
-      .pipe(tap((res) => {
-        this.form.patchValue({
-          background: res.background[0],
-          cardBackground: res.cardBackground[0]
-        }, { emitEvent: false });
-      }));
+  private getSurveyData(): any {
+    return this.surveyData$ = this.surveyService.getSurveyData()
   }
 
   private subscribeFormValueChanges(): void {
     this.form.valueChanges
-      .pipe(
-        debounceTime(500),
-        untilDestroyed(this)
-      )
+      .pipe(debounceTime(500), untilDestroyed(this))
       .subscribe((val) => {
-        this.questionData$.next({ questions: [val.questions[0]] });
+        this.questionData$.next({questions: [val.questions[0]]});
         this.cdr.detectChanges();
       })
   }
@@ -206,5 +229,24 @@ export class NewSurveyComponent implements OnInit, OnDestroy {
         this.tenantSettings = SettingsHttpAdapter.getTenantsSettings(res);
         this.cdr.detectChanges();
       });
+  }
+
+  private handleRouteParams(): Observable<any> {
+    return this.route.paramMap
+      .pipe(
+        untilDestroyed(this),
+        map((params: ParamMap) => params.get('id')
+        ),
+        tap(id => id), switchMap(id => {
+          if (id) {
+            return this.surveyService.getSurvey(id);
+          }
+          return of(null);
+        })
+      );
+  }
+
+  private getDefaultValue(queryData: any): any {
+    return NewSurveyForm.getDefaultValue(queryData);
   }
 }
